Be specific about DT or ACPI requirements
Be specific that an EBBR platform must provide either ACPI or Devicetree
data in the EFI configuration table, but not both. Platforms are allowed
to support both ACPI & DT booting as a configuration option, but only
one interface can be presented to the OS at a time.

Also add references to the ACPI and Devicetree specifications.

Fixes #5
Signed-off-by: Grant Likely <grant.likely@arm.com>
Reviewed-by: Tom Rini <trini@konsulko.com>
Acked-by: Mark Brown <broonie@kernel.org>