Move Scope into About This Document chapter
Scope doesn't need it's own chapter. Move it into the 'About This
Document' chapter. Also expand the text to place this document in
relation to the existing SBBR document. SBBR is the stricter of the two,
so EBBR can be considered a superset. (ie. all SBBR compliant platforms
are also EBBR compliant, but the converse is not true).

Signed-off-by: Grant Likely <grant.likely@arm.com>
Reviewed-by: Tom Rini <trini@konsulko.com>