Remove MP Start protocol from EBBR and revert to PSCI v1.0
EBBR platforms are unlikely to ever use the MP Start protocol used in
some AArch64 server systems. It was only used to support platforms
without EL3. No more EL2-only platforms are expected, and MP Start is
being removed from the next release of SBBR. Devicetree platforms in the
same scenario use the Linux spin table protocol instead.

MP Start was only included for completeness to align the SBBR and EBBR
specifications. Therefore remove it entirely, leaving PSCI and spin
table as the only acceptable implementations.

Also revert to PSCI v1.0 as that is the minimum level of functionality
required. PSCI v1.1 adds new capabilities, but is not required for
interoperability.

Suggested-by: Dong Wei <dong.wei@arm.com>
Signed-off-by: Grant Likely <grant.likely@arm.com>
Reviewed-by: Tom Rini <trini@konsulko.com>