Add requirements on runtime access to devices
To make it absolutely clear that runtime services cannot retain access
to devices that are 'owned' by the OS, add a section spelling the
situation out. Devices may be owned by the OS, or owned by runtime
services, but not both.

Also rework the RTC clock section to allow for the RTC being
inaccessible for exactly that reason.

Resolves: #1  "Need policy on sharing devices between FW and OS"
Resolves: #14  "RTC should be optional"

Signed-off-by: Grant Likely <grant.likely@arm.com>
Reviewed-by: Daniel Thompson <daniel.thompson@linaro.org>