DT: Mandate forward and backward compatibility
We currently do not mandate that a Devicetree stays compatible over
the lifetime of the OS or that it is compatible to an upstream OS
at all.

Fix that by restricting DT usage to only

  * upstream compatible
  * forwards and backwards compatible

device trees.

To indicate that, also explain the /ebbr node that can help people
validate whether a DT should and does fulfill the EBBR contract.

Signed-off-by: Alexander Graf <agraf@suse.de>