Simplify SetVariable() language
Instead of masking out GetVariable() when SetVariable() isn't available
during runtime services, simplify the requirements without losing the
ability to read variables by using the RuntimeServicesSupported variable
from UEFI v2.8 (in review for release); Mantis issue 1961.

Peter Jones's earlier patch also specified a Capsule-on-Disk format for
updating variables that the OS could store in the ESP. I've not included
that specification in this patch as it is logically a separate feature.
It may reappear in a separate patch at a later date, or it may get
proposed for inclusion in the UEFI spec proper.

Signed-off-by: Grant Likely <grant.likely@arm.com>
Cc: Peter Jones <pjones@redhat.com>