Remove unnecessary UEFI requirement tables
Nothing in the UEFI Requirements appendix is valuable.

The table of required boot services is unnecessary because it is an
exact duplicate of the UEFI boot services list in the UEFI spec (and it
also happens to be slightly incorrect) (UEFI 2.6.1). It is providing no
value to include in EBBR as all UEFI implementations are required to
implement the full set.

The tables of required core protocols are already specified in the UEFI
spec (UEFI 2.6.1)

The table of required media i/o protocols are already required if the
device supports booting from a disk device (UEFI 2.6.2).

The table of console protocols is similarly already required if a
console device is present.

It isn't clear that HII protocols need to be required. U-Boot does
implement them, but it doesn't appear to be a critical requirement on
whether or not an OSV can support the platform.

The tables of optional UEFI protocols isn't adding any value because it
doesn't require anything of implementers, and it doesn't provide any
commentary on when the protocols should be included. This is just
additional text.

Remove the lot to simplify the spec.

Signed-off-by: Grant Likely <grant.likely@arm.com>