Specify details of how a DTB is passed to the OS
Describe details of how to pass a DTB from firmware to the OS. Specifies
the GUID for passing the DTB via the EFI_SYSTEM_TABLE and the memory
type needs to be EfiACPIReclaimMemory. Also specifies that the DTB must
be handed off and no longer modified by firmware once execution of EFI
application and drivers begins.

Fixes: #45
Cc: Andrei Warkentin <awarkentin@vmware.com>
Cc: Francois Ozog <francois.ozog@linaro.org>
Cc: Ard Biesheuvel <ardb@kernel.org>
Signed-off-by: Grant Likely <grant.likely@arm.com>