Update spec to address UEFI 2.8 Errata A
UEFI 2.8 Errata A changed RuntimeServiceSupported value to be passed via
an EFI_RT_PROPERTIES_TABLE instead of via a runtime variable.

This is a breaking change and we should do a new EBBR release in short
order to keep EBBR up to date with what is in UEFI. However, real world
impact is minimal. U-Boot has already converted to the new method, and
Linux only ever implemented the EFI_RT_PROPERTIES_TABLE method.

Cc: Andrei Warkentin <awarkentin@vmware.com>
Cc: Francois Ozog <francois.ozog@linaro.org>
Cc: Ard Biesheuvel <ardb@kernel.org>
Signed-off-by: Grant Likely <grant.likely@arm.com>