Prepare for v1.0.1 point release
Minor changes compared to v1.0. Updated to match how UEFI spec
defines the RuntimeServicesSupported value and specify the UUID
to use when passing in a DTB. Remaining changes are editorial.

Updates the TravisCI key for building in github.com/glikely/ebbr.
Document build is handled in my personal fork because TravisCI is
not connected to the github.com/arm-software project.

Signed-off-by: Grant Likely <grant.likely@arm.com>