GetMemoryMap(), handling of no-map DT property
Closes: #52

The no-map property of the /reserved-memory DT node is used by Linux to
signal that a memory region shall not be mapped and that speculative access
shall not be permitted.

The memory map returned by GetMemoryMap() does not have a flag
corresponding to no-map. So the closest thing we can do is not to include
no-map reserved memory into the map returned by GetMemoryMap().

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>