GetMemoryMap(), handling of no-map DT property
Closes: #52

The no-map property of the /reserved-memory device tree nodes is used to
signal that a memory region shall not be accessed by the operating system,
even not via speculative access.

/reserved-memory nodes without the no-map property describe memory that
have special usage by the operating system.

This difference has to be reflected in the memory map returned by
GetMemoryMap().

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>