Change required services table titles to be more accurate
The table was using the headings "Boot Services" and "Runtime Services"
to describe which services need to be available both before and after
the call to ExitBootServices(). This is confusing because "Boot
Services" describes functions that are available only before the call to
ExitBootServices(), and "Runtime Services" are functions that may be
available both before and after ExitBootServices().

Change the table headings to not using confusing terms.
At the same time, reformat table to be easier to edit

Signed-off-by: Grant Likely <grant.likely@arm.com>