v2.0.1
This is a point update to EBBR spec.

This update fixes some typos and relaxes the requirement for
implementing ESRT. The ESRT relaxation is temporary due to released
software not yet being able to meet the requirement. Mainline U-Boot has
already fixed most of the issues so that ESRT can be reintroduced in the
next minor or major release.

This release also merges in requirements for RISC-V platforms which had
been discussed on the mailing list for some time now. It is safe to
merge in this point release as it doesn't affect requirements on other
architectures.

Signed-off-by: Grant Likely <grant.likely@arm.com>