Suggest conformance profile table for EBBRv2
The UEFI spec defines the conformance profile table.
EBBR v2.0 systems should publicize their compliance by having the
EFI_CONFOR EFI_CONFORMANCE_PROFILES_EBBR_2_0_GUID entry in the table.

We want the conformance profile table to be required for EBBR systems
but that can only be done at the next major EBBR release when the profile
GUID for that release is listed in the EBBR doc.

Signed-off-by: Jose Marinho <jose.marinho@arm.com>