File Format For Storing EFI Variables
Some UEFI enabled devices can only store EFI variables as a file on a block
device. This implies that at runtime the operating system must manage
changes to the EFI variable by updating the file.

Define a file-format for EFI variables that both the firmware and the
operating system can rely on.

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>