Add an extension for UEFI references
Add an extension python script to supplement our references to the UEFI
specification with corresponding section title and web-page hyperlink.

Hook the new extension into the `conf.py' sphinx configuration.

We keep the index under version control for caching, and we have a python
script to re-generate it.

Convert all the UEFI references to use the new extension and adapt a bit
for readability when necessary.

Mention the extension in the README.

While at it, tell git to ignore the folders produced by python when running
the extension.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>