CI: fix irc message computation
Commit messages with special characters such as "`" break the IRC message
computation because we retrieve them as "here documents", which makes the
shell choke.

Fix this by moving the IRC handling steps after checkout and by using git
commands to retrieve the commit message. As a bonus we can now retrieve the
hash, too.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>