trivial: adapt bibliography workaround
The bibliography chapter has some specific issues when rendered in pdf with
LaTeX, which results in duplicate chapter title (see github issue
sphinx-doc/sphinx#4775).

We have a workaround in place, which adds a chapter title only for the HTML
format, but this makes sphinx complain about the references having no title
for the non-HTML cases.

Change our workaround to follow the solution proposed in the aforementioned
github issue, which removes the sphinx warning.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>