Limit all citations labels to 8 characters
The Sphinx LaTeX writer has its maximum citation label length
MAX_CITATION_LABEL_LENGTH set to 8 characters. [1]
Larger citations labels cause misalignment in the resulting bibliography
for target latexpdf (which we track as issue #128).

As there is no easy way to change this limit, shorten all our citations
labels instead to a maximum of 8 characters and add a word about this
limitation in the README.

[1] https://github.com/sphinx-doc/sphinx/blob/master/sphinx/writers/latex.py

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>